pyproject.toml
[project]
name = "kmu-line-bot"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "yuka-04", email = "data2000analysis@gmail.com" }]
dependencies = [
    "line-bot-sdk>=3.11.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.111.1",
    "uvicorn>=0.30.3",
    "pandas>=2.0.3",
    "sentence_transformers>=3.0.1",
    "scikit-learn>=1.3.2",
    "openpyxl>=3.1.5",
]
# readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
include = ["main.py", "fqa_service.py", ".env", "pyproject.toml", "README.md"]
